Made SerializeParams() virtual, and added MakePresetFromChunk().
You can now using state chunks to:

a) Save a customized chunk of data (e.g. a hidden parameter, or an unsupported data type) in each preset. To do this you will need to override SerializeParams() and UnserializeParams(). In this case you can't use MakePreset() or MakePresetFromNamedParams() (because both require the preset to hold only the plug-in's "public" parameters in their predetermined order), but you can use the new MakePresetFromChunk(). When overriding SerializeParams() beware that it is called when the plug-in object is created, but before the GUI object is created, so if you need GetGUI() be sure to validate it first.

b) Save a customized chunk of data (e.g. a global setting like a GUI state variable) along with all presets (or along with one patch). In this case you need to override SerializePresets() and UnserializePresets(), which are called when a project is loaded, or when saving a preset bank. You will also need to override SerializeState() and UnserializeState(), which are called when loading/saving a single patch (e.g. from/to a .FXP file). Keep in mind that SerializePresets() / UnserializePresets() are only used in VST plug-ins, AU plug-ins always save/load only a single patch at the time using SerializeState() / UnserializeState().

Note: Previous to commit 8da3756 SerializeState() / UnserializeState() would be called, regardless of saving a single patch or a whole bank. Since commit 8da3756 SerializeState() / UnserializeState() is still called for a single patch, but SerializePresets() / UnserializePresets() is called for a whole bank. If your project relies on the old behaviour, then simply override SerializePresets() and make it call SerializeState().