IPlug: Made font cache methods static.
Made IGraphics::PrepDrawIText() and IGraphics::CacheFont() static, so you
can call IGraphics::PrepDrawIText() from an IControl's constructor (when
there is no IGraphics object yet) to preload a font.

Note that preloading fonts was already possible like this (warning: nasty
code up ahead!):

  IGraphics* pGUI = NULL;
  pGUI->PrepDrawIText(&mFont);

But this only works because PrepDrawIText() was effectively static already,
so then why not make it official (not to mention safe!).